fix_ppt_issue.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
修复PPT问题说明
"""

import os
from pathlib import Path

def explain_ppt_issue():
    """解释PPT问题并提供解决方案"""
    
    print("=== PPT生成问题分析 ===")
    print()
    
    # 检查文件
    base_path = Path(r"C:\Users\86151\Desktop\巡厂自动PPT")
    
    print("📋 问题诊断:")
    print("1. 之前的程序只是复制了模板文件")
    print("2. 没有实际创建多个幻灯片页面")
    print("3. 没有填充Excel数据到每个页面")
    print()
    
    print("🔧 真正需要做的事情:")
    print("✓ 读取Excel数据（31行记录）")
    print("✓ 为每行数据创建新的幻灯片")
    print("✓ 填充占位符内容")
    print("✓ 根据分类添加勾选标记")
    print("✓ 匹配并插入对应图片")
    print()
    
    # 检查数据
    excel_path = base_path / "Gemba巡厂_V2_20250920170854" / "Gemba巡厂_V2_20250920170854.xlsx"
    images_path = base_path / "Gemba巡厂_V2_20250920170854" / "Files" / "待整改--现场图片"
    
    print("📊 数据验证:")
    print(f"Excel文件: {'存在' if excel_path.exists() else '不存在'}")
    print(f"图片文件夹: {'存在' if images_path.exists() else '不存在'}")
    
    if images_path.exists():
        images = list(images_path.glob("*.jpeg"))
        print(f"图片数量: {len(images)} 张")
        
        # 显示几个示例数据和匹配的图片
        sample_problems = [
            "码垛机器人旁边漏雨",
            "成品库虚线还要有", 
            "板台里放了箱子，要分开",
            "主路不放木箱",
            "AGV会看该区域"
        ]
        
        print("\n🎯 图片匹配测试:")
        for problem in sample_problems:
            found = False
            for img in images:
                if problem in img.stem:
                    print(f"✓ '{problem}' -> {img.name}")
                    found = True
                    break
            if not found:
                print(f"✗ '{problem}' -> 未找到匹配")
    
    print()
    print("💡 解决方案:")
    print("1. 等待 python-pptx 库安装完成")
    print("2. 运行 real_ppt_generator.py")
    print("3. 这将创建真正的多页PPT（1首页 + 15数据页）")
    print()
    
    print("📈 预期结果:")
    print("- 第1页: 原始首页（日期已更新）")
    print("- 第2-16页: 每页显示一个问题的详细信息")
    print("- 每页包含: 问题区域、发现人、问题描述、分类勾选、现场图片")
    
    return True

def create_requirements_check():
    """检查依赖库状态"""
    print("\n=== 依赖库检查 ===")
    
    try:
        import pptx
        print("✓ python-pptx: 已安装")
        return True
    except ImportError:
        print("✗ python-pptx: 未安装")
        print("  请运行: pip install python-pptx")
        return False

def main():
    """主函数"""
    explain_ppt_issue()
    
    if create_requirements_check():
        print("\n🚀 准备就绪，可以生成真正的多页PPT!")
        print("运行命令: python real_ppt_generator.py")
    else:
        print("\n⏳ 等待依赖库安装完成...")
    
    return 0

if __name__ == "__main__":
    main()