requirements.txt
python-pptx==0.6.21
pandas==2.0.3
openpyxl==3.1.2
pathlib2==2.3.7